+++ firmware/src/drivers/grbl/nuts_bolts.c
@@ -83,14 +83,14 @@ uint8_t read_float(char *line, uint8_t *char_counter, float *float_ptr)
   // expected range of E0 to E-4.
   if (fval != 0) {
     while (exp <= -2) {
-      fval *= 0.01;
+      fval *= 0.01f;
       exp += 2;
     }
     if (exp < 0) {
-      fval *= 0.1;
+      fval *= 0.1f;
     } else if (exp > 0) {
       do {
-        fval *= 10.0;
+        fval *= 10.0f;
       } while (--exp > 0);
     }
   }
@@ -111,7 +111,7 @@ uint8_t read_float(char *line, uint8_t *char_counter, float *float_ptr)
 // Non-blocking delay function used for general operation and suspend features.
 void delay_sec(float seconds, uint8_t mode)
 {
- 	uint16_t i = ceil(1000/DWELL_TIME_STEP*seconds);
+ 	uint16_t i = (uint16_t)ceilf(1000/DWELL_TIME_STEP*seconds);
 	while (i-- > 0) {
 		if (sys.abort) { return; }
 		if (mode == DELAY_MODE_DWELL) {
@@ -158,20 +158,20 @@ void delay_us(uint32_t us)
 
 
 // Simple hypotenuse computation function.
-float hypot_f(float x, float y) { return(sqrt(x*x + y*y)); }
+float hypot_f(float x, float y) { return(sqrtf(x*x + y*y)); }
 
 
 float convert_delta_vector_to_unit_vector(float *vector)
 {
   uint8_t idx;
-  float magnitude = 0.0;
+  float magnitude = 0.0f;
   for (idx=0; idx<N_AXIS; idx++) {
-    if (vector[idx] != 0.0) {
+    if (vector[idx] != 0.0f) {
       magnitude += vector[idx]*vector[idx];
     }
   }
-  magnitude = sqrt(magnitude);
-  float inv_magnitude = 1.0/magnitude;
+  magnitude = sqrtf(magnitude);
+  float inv_magnitude = 1.0f/magnitude;
   for (idx=0; idx<N_AXIS; idx++) { vector[idx] *= inv_magnitude; }
   return(magnitude);
 }
@@ -183,7 +183,7 @@ float limit_value_by_axis_maximum(float *max_value, float *unit_vec)
   float limit_value = SOME_LARGE_VALUE;
   for (idx=0; idx<N_AXIS; idx++) {
     if (unit_vec[idx] != 0) {  // Avoid divide by zero.
-      limit_value = min(limit_value,fabs(max_value[idx]/unit_vec[idx]));
+      limit_value = min(limit_value,fabsf(max_value[idx]/unit_vec[idx]));
     }
   }
   return(limit_value);
